+++ frontend/src/api/models/index.ts
@@ -1,7 +1,6 @@
 export * from './k8s';
+export * from './kserve';
+export * from './kubeflow';
+export * from './kueue';
 export * from './odh';
 export * from './openShift';
-export * from './pipelines';
-export * from './kueue';
-export * from './trustyai';
-export * from './modelRegistry';
