+++ frontend/src/api/models/odh.ts
@@ -14,32 +14,11 @@ export const AcceleratorProfileModel: K8sModelCommon = {
   plural: 'acceleratorprofiles',
 };
 
-export const HardwareProfileModel = {
+export const HardwareProfileModel: K8sModelCommon = {
   apiVersion: 'v1alpha1',
   apiGroup: 'dashboard.opendatahub.io',
   kind: 'HardwareProfile',
   plural: 'hardwareprofiles',
-} satisfies K8sModelCommon;
-
-export const NotebookModel: K8sModelCommon = {
-  apiVersion: 'v1',
-  apiGroup: 'kubeflow.org',
-  kind: 'Notebook',
-  plural: 'notebooks',
-};
-
-export const ServingRuntimeModel: K8sModelCommon = {
-  apiVersion: 'v1alpha1',
-  apiGroup: 'serving.kserve.io',
-  kind: 'ServingRuntime',
-  plural: 'servingruntimes',
-};
-
-export const InferenceServiceModel: K8sModelCommon = {
-  apiVersion: 'v1beta1',
-  apiGroup: 'serving.kserve.io',
-  kind: 'InferenceService',
-  plural: 'inferenceservices',
 };
 
 export const NIMAccountModel: K8sModelCommon = {
@@ -54,4 +33,25 @@ export const AuthModel: K8sModelCommon = {
   apiGroup: 'services.platform.opendatahub.io',
   kind: 'Auth',
   plural: 'auths',
+};
+
+export const DataSciencePipelineApplicationModel: K8sModelCommon = {
+  apiVersion: 'v1',
+  apiGroup: 'datasciencepipelinesapplications.opendatahub.io',
+  kind: 'DataSciencePipelinesApplication',
+  plural: 'datasciencepipelinesapplications',
+};
+
+export const ModelRegistryModel: K8sModelCommon = {
+  apiVersion: 'v1alpha1',
+  apiGroup: 'modelregistry.opendatahub.io',
+  kind: 'ModelRegistry',
+  plural: 'modelregistries',
+};
+
+export const TrustyAIApplicationsModel: K8sModelCommon = {
+  apiVersion: 'v1alpha1',
+  apiGroup: 'trustyai.opendatahub.io',
+  kind: 'TrustyAIService',
+  plural: 'trustyaiservices',
 };
